Move v1 watchers to v1 namespace
