+++ watchers/v1/consul/consul_test.go
